+++ frontend/src/features/admin-form/create/common/components/BuilderSidebar/DrawerTabIcon.tsx
