+++ frontend/src/features/admin-form/create/common/components/CreatePageContent.tsx
@@ -1,9 +1,9 @@
 import {
   DrawerTabs,
   useCreatePageDrawer,
-} from '~features/admin-form-builder/CreatePageDrawerContext'
-import BuilderAndDesign from '~features/admin-form-builder/design'
-import CreatePageLogicTab from '~features/admin-form-builder/logic'
+} from '~features/admin-form/create/CreatePageDrawerContext'
+import BuilderAndDesign from '~features/admin-form/create/design'
+import CreatePageLogicTab from '~features/admin-form/create/logic'
 
 export const CreatePageContent = (): JSX.Element => {
   const { activeTab } = useCreatePageDrawer()
