+++ frontend/src/features/admin-form/create/design/components/BuilderDesignPlaceholder.tsx
